Add an additional appraisal mode, fixes and polish
Add the ability to appraise badges from the "Gym Badges" screen. (It
also works on the badges on player profile screen ;)
- There is reduced accuracy in this mode due to the progress bars being
shorter.
You can touch the badges on the screen shot after they've been appraised
to see the same stats that are given in the, default, single badge mode.
- The badge colour is guessed using the length the badge extends below
the progress bar, and what the most dominate colour of it is.
(We can be unsure of what colour the badge is due to how the badge is
positioned on the page)
This adds a nice new UI section to the top right, with two SVGs that
signal which mode you're using.

The new progress bar measuring code helps detecting pokemon overlapping
the right hand side of the progress bar. (This makes the result
"unsure")
We can also avoid an edge where we "detect" a progress bar on the
vertical scan, but it turns out to be just that 1 pixel wide, so we can
scan again.

JPEG handling code has not been improved.